Сделал event-controller
